requirements.txt
# 상세페이지 생성기 의존성
Pillow>=10.0.0
requests>=2.28.0
python-dotenv>=1.0.0
boto3>=1.34.0
